requirements.txt
accelerate==1.11.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
diffusers==0.35.2
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pillow==12.0.0
psutil==7.1.3
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
typer-slim==0.20.0
typing_extensions==4.15.0
urllib3==2.5.0
zipp==3.23.0
